Chore: createPost에 point, coupon 로직 분리 및 일부 수정 (#9)
