+++ src/main/java/com/example/docconneting/domain/coupon/entity/CouponHistory.java
@@ -1,6 +1,7 @@
 package com.example.docconneting.domain.coupon.entity;
 
 import jakarta.persistence.*;
+import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.CreatedDate;
