+++ src/main/java/com/example/docconneting/domain/coupon/repository/PatientCouponRepository.java
@@ -14,7 +14,6 @@ public interface PatientCouponRepository extends JpaRepository<PatientCoupon, Lo
     Optional<PatientCoupon> findPatientCouponByIdAndUserId(Long couponId, Long userId);
 
     boolean existsByUserIdAndCouponId(Long userId, Long couponId);
-    Optional<PatientCoupon> findByUserIdAndCouponId(Long userId, Long couponId);
 
     @Query("SELECT pc FROM PatientCoupon pc JOIN FETCH pc.coupon WHERE pc.user.id = :userId And pc.endDate > CURRENT_TIMESTAMP")
     Page<PatientCoupon> findAllByUserId(Pageable pageable, @Param("userId") Long userId);
